added message popup for errors
Moving text cursor over highlighted region will display the message associated with the error/note/warning.
It will also show the message in the status bar.
Automatic message showing is enabled by default. There are two new commands (Enable Auto Show Messages and Disable Auto Show Messages) to toggle this feature.
If it is disabled, then an option to show the message in a right-click context menu is available.